+++ src/app/constants/global.constants.ts
@@ -1,0 +1,3 @@
+export const LANGUAGES = {
+	esES: 'ES-ES',
+};